==> src/10_get_sentinel.R <==
#------------------------------------------------------------------------------
# Type: control script 
# Name: enviMet_simple_plants.R
# Author: Chris Reudenbach, creuden@gmail.com
# Description:  derives tree hulls and the corresponding values for 
#               tree species, height, LAD , albedo, transmissivity, root profile, 
#               lucc classes and seasonality
# Data: point cloudand dsm and dtm as derived by 10_CHM_Catalog.R  
#       sentinel 2 bands 2,3,4,8, classification of tree species  40_RS_high_resolution data
# Output: Envimet simple plant database file
# Copyright: Chris Reudenbach 2021, GPL (>= 3)
#------------------------------------------------------------------------------

## clean your environment
rm(list=ls()) 

# 0 - load packages
#-----------------------------


## dealing with the crs warnings is cumbersome and complex
## you may reduce the warnings with uncommenting  the following line
## for a deeper  however rarely less confusing understanding have a look at:
## https://rgdal.r-forge.r-project.org/articles/CRS_projections_transformations.html
## https://www.r-spatial.org/r/2020/03/17/wkt.html
rgdal::set_thin_PROJ6_warnings(TRUE)


# 1 - source files
#-----------------
source(file.path(envimaR::alternativeEnvi(root_folder = "~/edu/mpg-envinsys-plygrnd",
                                          alt_env_id = "COMPUTERNAME",
                                          alt_env_value = "PCRZP",
                                          alt_env_root_folder = "F:/BEN/edu/mpg-envinsys-plygrnd"),
                 "msc-phygeo-class-of-2020-creu/src/fun/000_setup.R"))

library(tidyr)
library(dplyr)
library(sen2r)
# 2 - define variables
#---------------------
out_dir =envrmt$path_sapflow
safe_dir  <- envrmt$path_SAFE  # folder to store downloaded SAFE
myextent  <- st_read(file.path(envrmt$path_sapflow,"sapflow_tree_segments_multichm_dalponte2016.gpkg"))

    
out_paths_1 <- sen2r(
  gui = FALSE,
  step_atmcorr = "l2a",
  online = FALSE,
  extent = myextent,
  extent_name = "sapflow",
  timewindow = c(as.Date("2021-06-12"), as.Date("2021-06-14")),
  list_prods = c("BOA","SCL"),
  list_indices = c("NDVI","MSAVI2"),
  list_rgb = c("RGB432B"),
  mask_type = "cloud_and_shadow",
  max_mask = 10, 
  path_l2a = safe_dir,
  server = "scihub",
  preprocess = TRUE,
  parallel = TRUE,
  sen2cor_use_dem = TRUE,
  max_cloud_safe = 10,
  overwrite = TRUE,
  path_out = 
    out_dir
)

fn=xfun::sans_ext(basename(list.files(paste0(envrmt$path_sapflow,"/BOA/"),pattern = "S2B2A")))

stack=raster::stack(paste0(envrmt$path_sapflow,"/BOA/",fn))
albedo = calculate_sentinel2_albedo(stack[[2]],
                                    stack[[3]],
                                    stack[[4]],
                                    stack[[8]])[[2]]
writeRaster(albedo,paste0(envrmt$path_sapflow,fn,"_albedo.tif"))




# sentinel LAI is a bit tricky and the easiest way is to utilise the webinterface  
# for running custom scripts and here LAI
# https://custom-scripts.sentinel-hub.com/sentinel-2/lai/
# preferable 
# https://apps.sentinel-hub.com/eo-browser/?zoom=15&lat=50.83534&lng=8.67551&themeId=DEFAULT-THEME&visualizationUrl=https%3A%2F%2Fservices.sentinel-hub.com%2Fogc%2Fwms%2F42924c6c-257a-4d04-9b8e-36387513a99c&evalscripturl=https%3A%2F%2Fraw.githubusercontent.com%2Fsentinel-hub%2Fcustom-scripts%2Fmaster%2Fsentinel-2%2Flai%2Fscript.js&datasetId=S2L1C&fromTime=2021-06-13T00%3A00%3A00.000Z&toTime=2021-06-13T23%3A59%3A59.999Z#custom-scrips
# in this case by default it is the following name
lai_fn= paste0(envrmt$path_sapflow,"2021-06-13-00:00_2021-06-13-23:59_Sentinel-2_L1C_Custom_script.tiff")


#https://raw.githubusercontent.com/sentinel-hub/custom-scripts/master/sentinel-2/lai/script.js
#https://apps.sentinel-hub.com/sentinel-playground/?source=S2&lat=50.83631315210404&lng=8.676238059997559&zoom=15&preset=CUSTOM&layers=B01,B02,B03&maxcc=0&gain=1.0&gamma=1.0&time=2020-12-01%7C2021-06-18&atmFilter=&showDates=false&evalscript=Ly9WRVJTSU9OPTMgKGF1dG8tY29udmVydGVkIGZyb20gMikKdmFyIGRlZ1RvUmFkID0gTWF0aC5QSSAvIDE4MDsKCmZ1bmN0aW9uIGV2YWx1YXRlUGl4ZWxPcmlnKHNhbXBsZXMpIHsKICB2YXIgc2FtcGxlID0gc2FtcGxlc1swXTsKICB2YXIgYjAzX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIwMywgMCwgMC4yNTMwNjE1MjA0NzE1NDIpOwogIHZhciBiMDRfbm9ybSA9IG5vcm1hbGl6ZShzYW1wbGUuQjA0LCAwLCAwLjI5MDM5MzU3NzkxMTMyOCk7CiAgdmFyIGIwNV9ub3JtID0gbm9ybWFsaXplKHNhbXBsZS5CMDUsIDAsIDAuMzA1Mzk4OTE1MjQ4NTU1KTsKICB2YXIgYjA2X25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIwNiwgMC4wMDY2Mzc5NzI1NDIyNTMsIDAuNjA4OTAwMzk1Nzk3ODg5KTsKICB2YXIgYjA3X25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIwNywgMC4wMTM5NzI3MjcwMTg5MzksIDAuNzUzODI3Mzg0MzIyOTI3KTsKICB2YXIgYjhhX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkI4QSwgMC4wMjY2OTAxMzgwODIwNjEsIDAuNzgyMDExNzcwNjY5MTc4KTsKICB2YXIgYjExX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIxMSwgMC4wMTYzODgwNzQxOTIyNTgsIDAuNDkzNzYxMzk3ODgzMDkyKTsKICB2YXIgYjEyX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIxMiwgMCwgMC40OTMwMjU5ODQ0NjAyMzEpOwogIHZhciB2aWV3WmVuX25vcm0gPSBub3JtYWxpemUoTWF0aC5jb3Moc2FtcGxlLnZpZXdaZW5pdGhNZWFuICogZGVnVG9SYWQpLCAwLjkxODU5NTQwMDU4MjA0NiwgMSk7CiAgdmFyIHN1blplbl9ub3JtICA9IG5vcm1hbGl6ZShNYXRoLmNvcyhzYW1wbGUuc3VuWmVuaXRoQW5nbGVzICogZGVnVG9SYWQpLCAwLjM0MjAyMjg3MTE1OTIwOCwgMC45MzYyMDY0MjkxNzU0MDIpOwogIHZhciByZWxBemltX25vcm0gPSBNYXRoLmNvcygoc2FtcGxlLnN1bkF6aW11dGhBbmdsZXMgLSBzYW1wbGUudmlld0F6aW11dGhNZWFuKSAqIGRlZ1RvUmFkKQoKICB2YXIgbjEgPSBuZXVyb24xKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjIgPSBuZXVyb24yKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjMgPSBuZXVyb24zKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjQgPSBuZXVyb240KGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjUgPSBuZXVyb241KGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKCiAgdmFyIGwyID0gbGF5ZXIyKG4xLCBuMiwgbjMsIG40LCBuNSk7CgogIHZhciBsYWkgPSBkZW5vcm1hbGl6ZShsMiwgMC4wMDAzMTkxODI1MzgzMDEsIDE0LjQ2NzUwOTQ1NDgxNTEpOwogIHJldHVybiB7CiAgICBkZWZhdWx0OiBbbGFpIC8gM10KICB9Cn0KCmZ1bmN0aW9uIG5ldXJvbjEoYjAzX25vcm0sYjA0X25vcm0sYjA1X25vcm0sYjA2X25vcm0sYjA3X25vcm0sYjhhX25vcm0sYjExX25vcm0sYjEyX25vcm0sIHZpZXdaZW5fbm9ybSxzdW5aZW5fbm9ybSxyZWxBemltX25vcm0pIHsKICB2YXIgc3VtID0KCSsgNC45NjIzODAzMDU1NTI3OQoJLSAwLjAyMzQwNjg3ODk2NjQ3MCAqIGIwM19ub3JtCgkrIDAuOTIxNjU1MTY0NjM2MzY2ICogYjA0X25vcm0KCSsgMC4xMzU1NzY1NDQwODAwOTkgKiBiMDVfbm9ybQoJLSAxLjkzODMzMTQ3MjM5Nzk1MCAqIGIwNl9ub3JtCgktIDMuMzQyNDk1ODE2MTIyNjgwICogYjA3X25vcm0KCSsgMC45MDIyNzc2NDgwMDk1NzYgKiBiOGFfbm9ybQoJKyAwLjIwNTM2MzUzODI1ODYxNCAqIGIxMV9ub3JtCgktIDAuMDQwNjA3ODQ0NzIxNzE2ICogYjEyX25vcm0KCS0gMC4wODMxOTY0MDk3MjcwOTIgKiB2aWV3WmVuX25vcm0KCSsgMC4yNjAwMjkyNzA3NzM4MDkgKiBzdW5aZW5fbm9ybQoJKyAwLjI4NDc2MTU2NzIxODg0NSAqIHJlbEF6aW1fbm9ybTsKCiAgcmV0dXJuIHRhbnNpZyhzdW0pOwp9CgpmdW5jdGlvbiBuZXVyb24yKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKSB7CiAgdmFyIHN1bSA9CgkrIDEuNDE2MDA4NDQzOTgxNTAwCgktIDAuMTMyNTU1NDgwODU2Njg0ICogYjAzX25vcm0KCS0gMC4xMzk1NzQ4MzczMzM1NDAgKiBiMDRfbm9ybQoJLSAxLjAxNDYwNjAxNjg5ODkyMCAqIGIwNV9ub3JtCgktIDEuMzMwODkwMDM4NjQ5MjcwICogYjA2X25vcm0KCSsgMC4wMzE3MzA2MjQ1MDMzNDEgKiBiMDdfbm9ybQoJLSAxLjQzMzU4MzU0MTMxNzA1MCAqIGI4YV9ub3JtCgktIDAuOTU5NjM3ODk4NTc0Njk5ICogYjExX25vcm0KCSsgMS4xMzMxMTU3MDY1NTEwMDAgKiBiMTJfbm9ybQoJKyAwLjIxNjYwMzg3NjU0MTYzMiAqIHZpZXdaZW5fbm9ybQoJKyAwLjQxMDY1MjMwMzc2MjgzOSAqIHN1blplbl9ub3JtCgkrIDAuMDY0NzYwMTU1NTQzNTA2ICogcmVsQXppbV9ub3JtOwoKICByZXR1cm4gdGFuc2lnKHN1bSk7Cn0KCmZ1bmN0aW9uIG5ldXJvbjMoYjAzX25vcm0sYjA0X25vcm0sYjA1X25vcm0sYjA2X25vcm0sYjA3X25vcm0sYjhhX25vcm0sYjExX25vcm0sYjEyX25vcm0sIHZpZXdaZW5fbm9ybSxzdW5aZW5fbm9ybSxyZWxBemltX25vcm0pIHsKICB2YXIgc3VtID0KCSsgMS4wNzU4OTcwNDcyMTMzMTAKCSsgMC4wODYwMTU5Nzc3MjQ4NjggKiBiMDNfbm9ybQoJKyAwLjYxNjY0ODc3Njg4MTQzNCAqIGIwNF9ub3JtCgkrIDAuNjc4MDAzODc2NDQ2NTU2ICogYjA1X25vcm0KCSsgMC4xNDExMDIzOTg2NDQ5NjggKiBiMDZfbm9ybQoJLSAwLjA5NjY4MjIwNjg4MzU0NiAqIGIwN19ub3JtCgktIDEuMTI4ODMyNjM4ODYyMjAwICogYjhhX25vcm0KCSsgMC4zMDIxODkxMDI3NDEzNzUgKiBiMTFfbm9ybQoJKyAwLjQzNDQ5NDkzNzI5OTcyNSAqIGIxMl9ub3JtCgktIDAuMDIxOTAzNjk5NDkwNTg5ICogdmlld1plbl9ub3JtCgktIDAuMjI4NDkyNDc2ODAyMjYzICogc3VuWmVuX25vcm0KCS0gMC4wMzk0NjA1Mzc1ODk4MjYgKiByZWxBemltX25vcm07CgogIHJldHVybiB0YW5zaWcoc3VtKTsKfQoKZnVuY3Rpb24gbmV1cm9uNChiMDNfbm9ybSxiMDRfbm9ybSxiMDVfbm9ybSxiMDZfbm9ybSxiMDdfbm9ybSxiOGFfbm9ybSxiMTFfbm9ybSxiMTJfbm9ybSwgdmlld1plbl9ub3JtLHN1blplbl9ub3JtLHJlbEF6aW1fbm9ybSkgewogIHZhciBzdW0gPQoJKyAxLjUzMzk4ODI2NDY1NTQyMAoJLSAwLjEwOTM2NjU5MzY3MDQwNCAqIGIwM19ub3JtCgktIDAuMDcxMDQ2MjYyOTcyNzI5ICogYjA0X25vcm0KCSsgMC4wNjQ1ODI0MTE0NzgzMjAgKiBiMDVfbm9ybQoJKyAyLjkwNjMyNTIzNjgyMzE2MCAqIGIwNl9ub3JtCgktIDAuNjczODczMTA4OTc5MTYzICogYjA3X25vcm0KCS0gMy44MzgwNTE4NjgyODA4NDAgKiBiOGFfbm9ybQoJKyAxLjY5NTk3OTM0NDUzMTUzMCAqIGIxMV9ub3JtCgkrIDAuMDQ2OTUwMjk2MDgxNzEzICogYjEyX25vcm0KCS0gMC4wNDk3MDk2NTI2ODgzNjUgKiB2aWV3WmVuX25vcm0KCSsgMC4wMjE4Mjk1NDU0MzA5OTQgKiBzdW5aZW5fbm9ybQoJKyAwLjA1NzQ4MzgyNzEwNDA5MSAqIHJlbEF6aW1fbm9ybTsKCiAgcmV0dXJuIHRhbnNpZyhzdW0pOwp9CgpmdW5jdGlvbiBuZXVyb241KGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKSB7CiAgdmFyIHN1bSA9CgkrIDMuMDI0MTE1OTMwNzU3MjMwCgktIDAuMDg5OTM5NDE2MTU5OTY5ICogYjAzX25vcm0KCSsgMC4xNzUzOTU0ODMxMDYxNDcgKiBiMDRfbm9ybQoJLSAwLjA4MTg0NzMyOTE3MjYyMCAqIGIwNV9ub3JtCgkrIDIuMjE5ODk1MzY3NDg3NzkwICogYjA2X25vcm0KCSsgMS43MTM4NzM5NzUxMzY4NTAgKiBiMDdfbm9ybQoJKyAwLjcxMzA2OTE4NjA5OTUzNCAqIGI4YV9ub3JtCgkrIDAuMTM4OTcwODEzNDk5MjAxICogYjExX25vcm0KCS0gMC4wNjA3NzE3NjE1MTgwMjUgKiBiMTJfbm9ybQoJKyAwLjEyNDI2MzM0MTI1NTQ3MyAqIHZpZXdaZW5fbm9ybQoJKyAwLjIxMDA4NjE0MDQwNDM1MSAqIHN1blplbl9ub3JtCgktIDAuMTgzODc4MTM4NzAwMzQxICogcmVsQXppbV9ub3JtOwoKICByZXR1cm4gdGFuc2lnKHN1bSk7Cn0KCmZ1bmN0aW9uIGxheWVyMihuZXVyb24xLCBuZXVyb24yLCBuZXVyb24zLCBuZXVyb240LCBuZXVyb241KSB7CiAgdmFyIHN1bSA9CgkrIDEuMDk2OTYzMTA3MDc3MjIwCgktIDEuNTAwMTM1NDg5NzI4NzMwICogbmV1cm9uMQoJLSAwLjA5NjI4MzI2OTEyMTUwMyAqIG5ldXJvbjIKCS0gMC4xOTQ5MzU5MzA1NzcwOTQgKiBuZXVyb24zCgktIDAuMzUyMzA1ODk1NzU1NTkxICogbmV1cm9uNAoJKyAwLjA3NTEwNzQxNTg0NzQ3MyAqIG5ldXJvbjU7CgogIHJldHVybiBzdW07Cn0KCmZ1bmN0aW9uIG5vcm1hbGl6ZSh1bm5vcm1hbGl6ZWQsIG1pbiwgbWF4KSB7CiAgcmV0dXJuIDIgKiAodW5ub3JtYWxpemVkIC0gbWluKSAvIChtYXggLSBtaW4pIC0gMTsKfQpmdW5jdGlvbiBkZW5vcm1hbGl6ZShub3JtYWxpemVkLCBtaW4sIG1heCkgewogIHJldHVybiAwLjUgKiAobm9ybWFsaXplZCArIDEpICogKG1heCAtIG1pbikgKyBtaW47Cn0KZnVuY3Rpb24gdGFuc2lnKGlucHV0KSB7CiAgcmV0dXJuIDIgLyAoMSArIE1hdGguZXhwKC0yICogaW5wdXQpKSAtIDE7IAp9CgpmdW5jdGlvbiBzZXR1cCgpIHsKICByZXR1cm4gewogICAgaW5wdXQ6IFt7CiAgICAgIGJhbmRzOiBbCiAgICAgICAgICAiQjAzIiwKICAgICAgICAgICJCMDQiLAogICAgICAgICAgIkIwNSIsCiAgICAgICAgICAiQjA2IiwKICAgICAgICAgICJCMDciLAogICAgICAgICAgIkI4QSIsCiAgICAgICAgICAiQjExIiwKICAgICAgICAgICJCMTIiLAogICAgICAgICAgInZpZXdaZW5pdGhNZWFuIiwKICAgICAgICAgICJ2aWV3QXppbXV0aE1lYW4iLAogICAgICAgICAgInN1blplbml0aEFuZ2xlcyIsCiAgICAgICAgICAic3VuQXppbXV0aEFuZ2xlcyIKICAgICAgXQogICAgfV0sCiAgICBvdXRwdXQ6IFsKICAgICAgICB7CiAgICAgICAgICBpZDogImRlZmF1bHQiLAogICAgICAgICAgc2FtcGxlVHlwZTogIkFVVE8iLAogICAgICAgICAgYmFuZHM6IDEKICAgICAgICB9CiAgICBdCiAgfQp9CgpmdW5jdGlvbiBldmFsdWF0ZVBpeGVsKHNhbXBsZSwgc2NlbmUsIG1ldGFkYXRhLCBjdXN0b21EYXRhLCBvdXRwdXRNZXRhZGF0YSkgewogIGNvbnN0IHJlc3VsdCA9IGV2YWx1YXRlUGl4ZWxPcmlnKFtzYW1wbGVdLCBbc2NlbmVdLCBtZXRhZGF0YSwgY3VzdG9tRGF0YSwgb3V0cHV0TWV0YWRhdGEpOwogIHJldHVybiByZXN1bHRbT2JqZWN0LmtleXMocmVzdWx0KVswXV07Cn0K&evalscripturl=https://raw.githubusercontent.com/sentinel-hub/custom-scripts/master/sentinel-2/lai/script.js&showImage
#https://apps.sentinel-hub.com/eo-browser/?zoom=15&lat=50.83534&lng=8.67551&themeId=DEFAULT-THEME&visualizationUrl=https%3A%2F%2Fservices.sentinel-hub.com%2Fogc%2Fwms%2F42924c6c-257a-4d04-9b8e-36387513a99c&evalscripturl=https%3A%2F%2Fraw.githubusercontent.com%2Fsentinel-hub%2Fcustom-scripts%2Fmaster%2Fsentinel-2%2Flai%2Fscript.js&datasetId=S2L1C&fromTime=2021-06-13T00%3A00%3A00.000Z&toTime=2021-06-13T23%3A59%3A59.999Z#custom-script
#https://apps.sentinel-hub.com/eo-browser/?zoom=15&lat=50.83534&lng=8.67551&themeId=DEFAULT-THEME&visualizationUrl=https%3A%2F%2Fservices.sentinel-hub.com%2Fogc%2Fwms%2F42924c6c-257a-4d04-9b8e-36387513a99c&evalscripturl=https%3A%2F%2Fraw.githubusercontent.com%2Fsentinel-hub%2Fcustom-scripts%2Fmaster%2Fsentinel-2%2Flai%2Fscript.js&datasetId=S2L1C&fromTime=2021-06-13T00%3A00%3A00.000Z&toTime=2021-06-13T23%3A59%3A59.999Z#custom-script   
# poorer in performance
#https://apps.sentinel-hub.com/sentinel-playground/?source=S2&lat=50.83631315210404&lng=8.676238059997559&zoom=15&preset=CUSTOM&layers=B01,B02,B03&maxcc=0&gain=1.0&gamma=1.0&time=2020-12-01%7C2021-06-13&atmFilter=&showDates=false&evalscript=Ly9WRVJTSU9OPTMgKGF1dG8tY29udmVydGVkIGZyb20gMikKdmFyIGRlZ1RvUmFkID0gTWF0aC5QSSAvIDE4MDsKCmZ1bmN0aW9uIGV2YWx1YXRlUGl4ZWxPcmlnKHNhbXBsZXMpIHsKICB2YXIgc2FtcGxlID0gc2FtcGxlc1swXTsKICB2YXIgYjAzX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIwMywgMCwgMC4yNTMwNjE1MjA0NzE1NDIpOwogIHZhciBiMDRfbm9ybSA9IG5vcm1hbGl6ZShzYW1wbGUuQjA0LCAwLCAwLjI5MDM5MzU3NzkxMTMyOCk7CiAgdmFyIGIwNV9ub3JtID0gbm9ybWFsaXplKHNhbXBsZS5CMDUsIDAsIDAuMzA1Mzk4OTE1MjQ4NTU1KTsKICB2YXIgYjA2X25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIwNiwgMC4wMDY2Mzc5NzI1NDIyNTMsIDAuNjA4OTAwMzk1Nzk3ODg5KTsKICB2YXIgYjA3X25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIwNywgMC4wMTM5NzI3MjcwMTg5MzksIDAuNzUzODI3Mzg0MzIyOTI3KTsKICB2YXIgYjhhX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkI4QSwgMC4wMjY2OTAxMzgwODIwNjEsIDAuNzgyMDExNzcwNjY5MTc4KTsKICB2YXIgYjExX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIxMSwgMC4wMTYzODgwNzQxOTIyNTgsIDAuNDkzNzYxMzk3ODgzMDkyKTsKICB2YXIgYjEyX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIxMiwgMCwgMC40OTMwMjU5ODQ0NjAyMzEpOwogIHZhciB2aWV3WmVuX25vcm0gPSBub3JtYWxpemUoTWF0aC5jb3Moc2FtcGxlLnZpZXdaZW5pdGhNZWFuICogZGVnVG9SYWQpLCAwLjkxODU5NTQwMDU4MjA0NiwgMSk7CiAgdmFyIHN1blplbl9ub3JtICA9IG5vcm1hbGl6ZShNYXRoLmNvcyhzYW1wbGUuc3VuWmVuaXRoQW5nbGVzICogZGVnVG9SYWQpLCAwLjM0MjAyMjg3MTE1OTIwOCwgMC45MzYyMDY0MjkxNzU0MDIpOwogIHZhciByZWxBemltX25vcm0gPSBNYXRoLmNvcygoc2FtcGxlLnN1bkF6aW11dGhBbmdsZXMgLSBzYW1wbGUudmlld0F6aW11dGhNZWFuKSAqIGRlZ1RvUmFkKQoKICB2YXIgbjEgPSBuZXVyb24xKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjIgPSBuZXVyb24yKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjMgPSBuZXVyb24zKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjQgPSBuZXVyb240KGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjUgPSBuZXVyb241KGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKCiAgdmFyIGwyID0gbGF5ZXIyKG4xLCBuMiwgbjMsIG40LCBuNSk7CgogIHZhciBsYWkgPSBkZW5vcm1hbGl6ZShsMiwgMC4wMDAzMTkxODI1MzgzMDEsIDE0LjQ2NzUwOTQ1NDgxNTEpOwogIHJldHVybiB7CiAgICBkZWZhdWx0OiBbbGFpIC8gM10KICB9Cn0KCmZ1bmN0aW9uIG5ldXJvbjEoYjAzX25vcm0sYjA0X25vcm0sYjA1X25vcm0sYjA2X25vcm0sYjA3X25vcm0sYjhhX25vcm0sYjExX25vcm0sYjEyX25vcm0sIHZpZXdaZW5fbm9ybSxzdW5aZW5fbm9ybSxyZWxBemltX25vcm0pIHsKICB2YXIgc3VtID0KCSsgNC45NjIzODAzMDU1NTI3OQoJLSAwLjAyMzQwNjg3ODk2NjQ3MCAqIGIwM19ub3JtCgkrIDAuOTIxNjU1MTY0NjM2MzY2ICogYjA0X25vcm0KCSsgMC4xMzU1NzY1NDQwODAwOTkgKiBiMDVfbm9ybQoJLSAxLjkzODMzMTQ3MjM5Nzk1MCAqIGIwNl9ub3JtCgktIDMuMzQyNDk1ODE2MTIyNjgwICogYjA3X25vcm0KCSsgMC45MDIyNzc2NDgwMDk1NzYgKiBiOGFfbm9ybQoJKyAwLjIwNTM2MzUzODI1ODYxNCAqIGIxMV9ub3JtCgktIDAuMDQwNjA3ODQ0NzIxNzE2ICogYjEyX25vcm0KCS0gMC4wODMxOTY0MDk3MjcwOTIgKiB2aWV3WmVuX25vcm0KCSsgMC4yNjAwMjkyNzA3NzM4MDkgKiBzdW5aZW5fbm9ybQoJKyAwLjI4NDc2MTU2NzIxODg0NSAqIHJlbEF6aW1fbm9ybTsKCiAgcmV0dXJuIHRhbnNpZyhzdW0pOwp9CgpmdW5jdGlvbiBuZXVyb24yKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKSB7CiAgdmFyIHN1bSA9CgkrIDEuNDE2MDA4NDQzOTgxNTAwCgktIDAuMTMyNTU1NDgwODU2Njg0ICogYjAzX25vcm0KCS0gMC4xMzk1NzQ4MzczMzM1NDAgKiBiMDRfbm9ybQoJLSAxLjAxNDYwNjAxNjg5ODkyMCAqIGIwNV9ub3JtCgktIDEuMzMwODkwMDM4NjQ5MjcwICogYjA2X25vcm0KCSsgMC4wMzE3MzA2MjQ1MDMzNDEgKiBiMDdfbm9ybQoJLSAxLjQzMzU4MzU0MTMxNzA1MCAqIGI4YV9ub3JtCgktIDAuOTU5NjM3ODk4NTc0Njk5ICogYjExX25vcm0KCSsgMS4xMzMxMTU3MDY1NTEwMDAgKiBiMTJfbm9ybQoJKyAwLjIxNjYwMzg3NjU0MTYzMiAqIHZpZXdaZW5fbm9ybQoJKyAwLjQxMDY1MjMwMzc2MjgzOSAqIHN1blplbl9ub3JtCgkrIDAuMDY0NzYwMTU1NTQzNTA2ICogcmVsQXppbV9ub3JtOwoKICByZXR1cm4gdGFuc2lnKHN1bSk7Cn0KCmZ1bmN0aW9uIG5ldXJvbjMoYjAzX25vcm0sYjA0X25vcm0sYjA1X25vcm0sYjA2X25vcm0sYjA3X25vcm0sYjhhX25vcm0sYjExX25vcm0sYjEyX25vcm0sIHZpZXdaZW5fbm9ybSxzdW5aZW5fbm9ybSxyZWxBemltX25vcm0pIHsKICB2YXIgc3VtID0KCSsgMS4wNzU4OTcwNDcyMTMzMTAKCSsgMC4wODYwMTU5Nzc3MjQ4NjggKiBiMDNfbm9ybQoJKyAwLjYxNjY0ODc3Njg4MTQzNCAqIGIwNF9ub3JtCgkrIDAuNjc4MDAzODc2NDQ2NTU2ICogYjA1X25vcm0KCSsgMC4xNDExMDIzOTg2NDQ5NjggKiBiMDZfbm9ybQoJLSAwLjA5NjY4MjIwNjg4MzU0NiAqIGIwN19ub3JtCgktIDEuMTI4ODMyNjM4ODYyMjAwICogYjhhX25vcm0KCSsgMC4zMDIxODkxMDI3NDEzNzUgKiBiMTFfbm9ybQoJKyAwLjQzNDQ5NDkzNzI5OTcyNSAqIGIxMl9ub3JtCgktIDAuMDIxOTAzNjk5NDkwNTg5ICogdmlld1plbl9ub3JtCgktIDAuMjI4NDkyNDc2ODAyMjYzICogc3VuWmVuX25vcm0KCS0gMC4wMzk0NjA1Mzc1ODk4MjYgKiByZWxBemltX25vcm07CgogIHJldHVybiB0YW5zaWcoc3VtKTsKfQoKZnVuY3Rpb24gbmV1cm9uNChiMDNfbm9ybSxiMDRfbm9ybSxiMDVfbm9ybSxiMDZfbm9ybSxiMDdfbm9ybSxiOGFfbm9ybSxiMTFfbm9ybSxiMTJfbm9ybSwgdmlld1plbl9ub3JtLHN1blplbl9ub3JtLHJlbEF6aW1fbm9ybSkgewogIHZhciBzdW0gPQoJKyAxLjUzMzk4ODI2NDY1NTQyMAoJLSAwLjEwOTM2NjU5MzY3MDQwNCAqIGIwM19ub3JtCgktIDAuMDcxMDQ2MjYyOTcyNzI5ICogYjA0X25vcm0KCSsgMC4wNjQ1ODI0MTE0NzgzMjAgKiBiMDVfbm9ybQoJKyAyLjkwNjMyNTIzNjgyMzE2MCAqIGIwNl9ub3JtCgktIDAuNjczODczMTA4OTc5MTYzICogYjA3X25vcm0KCS0gMy44MzgwNTE4NjgyODA4NDAgKiBiOGFfbm9ybQoJKyAxLjY5NTk3OTM0NDUzMTUzMCAqIGIxMV9ub3JtCgkrIDAuMDQ2OTUwMjk2MDgxNzEzICogYjEyX25vcm0KCS0gMC4wNDk3MDk2NTI2ODgzNjUgKiB2aWV3WmVuX25vcm0KCSsgMC4wMjE4Mjk1NDU0MzA5OTQgKiBzdW5aZW5fbm9ybQoJKyAwLjA1NzQ4MzgyNzEwNDA5MSAqIHJlbEF6aW1fbm9ybTsKCiAgcmV0dXJuIHRhbnNpZyhzdW0pOwp9CgpmdW5jdGlvbiBuZXVyb241KGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKSB7CiAgdmFyIHN1bSA9CgkrIDMuMDI0MTE1OTMwNzU3MjMwCgktIDAuMDg5OTM5NDE2MTU5OTY5ICogYjAzX25vcm0KCSsgMC4xNzUzOTU0ODMxMDYxNDcgKiBiMDRfbm9ybQoJLSAwLjA4MTg0NzMyOTE3MjYyMCAqIGIwNV9ub3JtCgkrIDIuMjE5ODk1MzY3NDg3NzkwICogYjA2X25vcm0KCSsgMS43MTM4NzM5NzUxMzY4NTAgKiBiMDdfbm9ybQoJKyAwLjcxMzA2OTE4NjA5OTUzNCAqIGI4YV9ub3JtCgkrIDAuMTM4OTcwODEzNDk5MjAxICogYjExX25vcm0KCS0gMC4wNjA3NzE3NjE1MTgwMjUgKiBiMTJfbm9ybQoJKyAwLjEyNDI2MzM0MTI1NTQ3MyAqIHZpZXdaZW5fbm9ybQoJKyAwLjIxMDA4NjE0MDQwNDM1MSAqIHN1blplbl9ub3JtCgktIDAuMTgzODc4MTM4NzAwMzQxICogcmVsQXppbV9ub3JtOwoKICByZXR1cm4gdGFuc2lnKHN1bSk7Cn0KCmZ1bmN0aW9uIGxheWVyMihuZXVyb24xLCBuZXVyb24yLCBuZXVyb24zLCBuZXVyb240LCBuZXVyb241KSB7CiAgdmFyIHN1bSA9CgkrIDEuMDk2OTYzMTA3MDc3MjIwCgktIDEuNTAwMTM1NDg5NzI4NzMwICogbmV1cm9uMQoJLSAwLjA5NjI4MzI2OTEyMTUwMyAqIG5ldXJvbjIKCS0gMC4xOTQ5MzU5MzA1NzcwOTQgKiBuZXVyb24zCgktIDAuMzUyMzA1ODk1NzU1NTkxICogbmV1cm9uNAoJKyAwLjA3NTEwNzQxNTg0NzQ3MyAqIG5ldXJvbjU7CgogIHJldHVybiBzdW07Cn0KCmZ1bmN0aW9uIG5vcm1hbGl6ZSh1bm5vcm1hbGl6ZWQsIG1pbiwgbWF4KSB7CiAgcmV0dXJuIDIgKiAodW5ub3JtYWxpemVkIC0gbWluKSAvIChtYXggLSBtaW4pIC0gMTsKfQpmdW5jdGlvbiBkZW5vcm1hbGl6ZShub3JtYWxpemVkLCBtaW4sIG1heCkgewogIHJldHVybiAwLjUgKiAobm9ybWFsaXplZCArIDEpICogKG1heCAtIG1pbikgKyBtaW47Cn0KZnVuY3Rpb24gdGFuc2lnKGlucHV0KSB7CiAgcmV0dXJuIDIgLyAoMSArIE1hdGguZXhwKC0yICogaW5wdXQpKSAtIDE7IAp9CgpmdW5jdGlvbiBzZXR1cCgpIHsKICByZXR1cm4gewogICAgaW5wdXQ6IFt7CiAgICAgIGJhbmRzOiBbCiAgICAgICAgICAiQjAzIiwKICAgICAgICAgICJCMDQiLAogICAgICAgICAgIkIwNSIsCiAgICAgICAgICAiQjA2IiwKICAgICAgICAgICJCMDciLAogICAgICAgICAgIkI4QSIsCiAgICAgICAgICAiQjExIiwKICAgICAgICAgICJCMTIiLAogICAgICAgICAgInZpZXdaZW5pdGhNZWFuIiwKICAgICAgICAgICJ2aWV3QXppbXV0aE1lYW4iLAogICAgICAgICAgInN1blplbml0aEFuZ2xlcyIsCiAgICAgICAgICAic3VuQXppbXV0aEFuZ2xlcyIKICAgICAgXQogICAgfV0sCiAgICBvdXRwdXQ6IFsKICAgICAgICB7CiAgICAgICAgICBpZDogImRlZmF1bHQiLAogICAgICAgICAgc2FtcGxlVHlwZTogIkFVVE8iLAogICAgICAgICAgYmFuZHM6IDEKICAgICAgICB9CiAgICBdCiAgfQp9CgpmdW5jdGlvbiBldmFsdWF0ZVBpeGVsKHNhbXBsZSwgc2NlbmUsIG1ldGFkYXRhLCBjdXN0b21EYXRhLCBvdXRwdXRNZXRhZGF0YSkgewogIGNvbnN0IHJlc3VsdCA9IGV2YWx1YXRlUGl4ZWxPcmlnKFtzYW1wbGVdLCBbc2NlbmVdLCBtZXRhZGF0YSwgY3VzdG9tRGF0YSwgb3V0cHV0TWV0YWRhdGEpOwogIHJldHVybiByZXN1bHRbT2JqZWN0LmtleXMocmVzdWx0KVswXV07Cn0K&evalscripturl=https://raw.githubusercontent.com/sentinel-hub/custom-scripts/master/sentinel-2/lai/script.js&showImage
#https://apps.sentinel-hub.com/sentinel-playground/?source=S2&lat=50.83631315210404&lng=8.676238059997559&zoom=15&preset=CUSTOM&layers=B01,B02,B03&maxcc=0&gain=1.0&gamma=1.0&time=2020-12-01%7C2021-06-13&atmFilter=&showDates=false&evalscript=Ly9WRVJTSU9OPTMgKGF1dG8tY29udmVydGVkIGZyb20gMikKdmFyIGRlZ1RvUmFkID0gTWF0aC5QSSAvIDE4MDsKCmZ1bmN0aW9uIGV2YWx1YXRlUGl4ZWxPcmlnKHNhbXBsZXMpIHsKICB2YXIgc2FtcGxlID0gc2FtcGxlc1swXTsKICB2YXIgYjAzX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIwMywgMCwgMC4yNTMwNjE1MjA0NzE1NDIpOwogIHZhciBiMDRfbm9ybSA9IG5vcm1hbGl6ZShzYW1wbGUuQjA0LCAwLCAwLjI5MDM5MzU3NzkxMTMyOCk7CiAgdmFyIGIwNV9ub3JtID0gbm9ybWFsaXplKHNhbXBsZS5CMDUsIDAsIDAuMzA1Mzk4OTE1MjQ4NTU1KTsKICB2YXIgYjA2X25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIwNiwgMC4wMDY2Mzc5NzI1NDIyNTMsIDAuNjA4OTAwMzk1Nzk3ODg5KTsKICB2YXIgYjA3X25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIwNywgMC4wMTM5NzI3MjcwMTg5MzksIDAuNzUzODI3Mzg0MzIyOTI3KTsKICB2YXIgYjhhX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkI4QSwgMC4wMjY2OTAxMzgwODIwNjEsIDAuNzgyMDExNzcwNjY5MTc4KTsKICB2YXIgYjExX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIxMSwgMC4wMTYzODgwNzQxOTIyNTgsIDAuNDkzNzYxMzk3ODgzMDkyKTsKICB2YXIgYjEyX25vcm0gPSBub3JtYWxpemUoc2FtcGxlLkIxMiwgMCwgMC40OTMwMjU5ODQ0NjAyMzEpOwogIHZhciB2aWV3WmVuX25vcm0gPSBub3JtYWxpemUoTWF0aC5jb3Moc2FtcGxlLnZpZXdaZW5pdGhNZWFuICogZGVnVG9SYWQpLCAwLjkxODU5NTQwMDU4MjA0NiwgMSk7CiAgdmFyIHN1blplbl9ub3JtICA9IG5vcm1hbGl6ZShNYXRoLmNvcyhzYW1wbGUuc3VuWmVuaXRoQW5nbGVzICogZGVnVG9SYWQpLCAwLjM0MjAyMjg3MTE1OTIwOCwgMC45MzYyMDY0MjkxNzU0MDIpOwogIHZhciByZWxBemltX25vcm0gPSBNYXRoLmNvcygoc2FtcGxlLnN1bkF6aW11dGhBbmdsZXMgLSBzYW1wbGUudmlld0F6aW11dGhNZWFuKSAqIGRlZ1RvUmFkKQoKICB2YXIgbjEgPSBuZXVyb24xKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjIgPSBuZXVyb24yKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjMgPSBuZXVyb24zKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjQgPSBuZXVyb240KGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKICB2YXIgbjUgPSBuZXVyb241KGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKTsKCiAgdmFyIGwyID0gbGF5ZXIyKG4xLCBuMiwgbjMsIG40LCBuNSk7CgogIHZhciBsYWkgPSBkZW5vcm1hbGl6ZShsMiwgMC4wMDAzMTkxODI1MzgzMDEsIDE0LjQ2NzUwOTQ1NDgxNTEpOwogIHJldHVybiB7CiAgICBkZWZhdWx0OiBbbGFpIC8gM10KICB9Cn0KCmZ1bmN0aW9uIG5ldXJvbjEoYjAzX25vcm0sYjA0X25vcm0sYjA1X25vcm0sYjA2X25vcm0sYjA3X25vcm0sYjhhX25vcm0sYjExX25vcm0sYjEyX25vcm0sIHZpZXdaZW5fbm9ybSxzdW5aZW5fbm9ybSxyZWxBemltX25vcm0pIHsKICB2YXIgc3VtID0KCSsgNC45NjIzODAzMDU1NTI3OQoJLSAwLjAyMzQwNjg3ODk2NjQ3MCAqIGIwM19ub3JtCgkrIDAuOTIxNjU1MTY0NjM2MzY2ICogYjA0X25vcm0KCSsgMC4xMzU1NzY1NDQwODAwOTkgKiBiMDVfbm9ybQoJLSAxLjkzODMzMTQ3MjM5Nzk1MCAqIGIwNl9ub3JtCgktIDMuMzQyNDk1ODE2MTIyNjgwICogYjA3X25vcm0KCSsgMC45MDIyNzc2NDgwMDk1NzYgKiBiOGFfbm9ybQoJKyAwLjIwNTM2MzUzODI1ODYxNCAqIGIxMV9ub3JtCgktIDAuMDQwNjA3ODQ0NzIxNzE2ICogYjEyX25vcm0KCS0gMC4wODMxOTY0MDk3MjcwOTIgKiB2aWV3WmVuX25vcm0KCSsgMC4yNjAwMjkyNzA3NzM4MDkgKiBzdW5aZW5fbm9ybQoJKyAwLjI4NDc2MTU2NzIxODg0NSAqIHJlbEF6aW1fbm9ybTsKCiAgcmV0dXJuIHRhbnNpZyhzdW0pOwp9CgpmdW5jdGlvbiBuZXVyb24yKGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKSB7CiAgdmFyIHN1bSA9CgkrIDEuNDE2MDA4NDQzOTgxNTAwCgktIDAuMTMyNTU1NDgwODU2Njg0ICogYjAzX25vcm0KCS0gMC4xMzk1NzQ4MzczMzM1NDAgKiBiMDRfbm9ybQoJLSAxLjAxNDYwNjAxNjg5ODkyMCAqIGIwNV9ub3JtCgktIDEuMzMwODkwMDM4NjQ5MjcwICogYjA2X25vcm0KCSsgMC4wMzE3MzA2MjQ1MDMzNDEgKiBiMDdfbm9ybQoJLSAxLjQzMzU4MzU0MTMxNzA1MCAqIGI4YV9ub3JtCgktIDAuOTU5NjM3ODk4NTc0Njk5ICogYjExX25vcm0KCSsgMS4xMzMxMTU3MDY1NTEwMDAgKiBiMTJfbm9ybQoJKyAwLjIxNjYwMzg3NjU0MTYzMiAqIHZpZXdaZW5fbm9ybQoJKyAwLjQxMDY1MjMwMzc2MjgzOSAqIHN1blplbl9ub3JtCgkrIDAuMDY0NzYwMTU1NTQzNTA2ICogcmVsQXppbV9ub3JtOwoKICByZXR1cm4gdGFuc2lnKHN1bSk7Cn0KCmZ1bmN0aW9uIG5ldXJvbjMoYjAzX25vcm0sYjA0X25vcm0sYjA1X25vcm0sYjA2X25vcm0sYjA3X25vcm0sYjhhX25vcm0sYjExX25vcm0sYjEyX25vcm0sIHZpZXdaZW5fbm9ybSxzdW5aZW5fbm9ybSxyZWxBemltX25vcm0pIHsKICB2YXIgc3VtID0KCSsgMS4wNzU4OTcwNDcyMTMzMTAKCSsgMC4wODYwMTU5Nzc3MjQ4NjggKiBiMDNfbm9ybQoJKyAwLjYxNjY0ODc3Njg4MTQzNCAqIGIwNF9ub3JtCgkrIDAuNjc4MDAzODc2NDQ2NTU2ICogYjA1X25vcm0KCSsgMC4xNDExMDIzOTg2NDQ5NjggKiBiMDZfbm9ybQoJLSAwLjA5NjY4MjIwNjg4MzU0NiAqIGIwN19ub3JtCgktIDEuMTI4ODMyNjM4ODYyMjAwICogYjhhX25vcm0KCSsgMC4zMDIxODkxMDI3NDEzNzUgKiBiMTFfbm9ybQoJKyAwLjQzNDQ5NDkzNzI5OTcyNSAqIGIxMl9ub3JtCgktIDAuMDIxOTAzNjk5NDkwNTg5ICogdmlld1plbl9ub3JtCgktIDAuMjI4NDkyNDc2ODAyMjYzICogc3VuWmVuX25vcm0KCS0gMC4wMzk0NjA1Mzc1ODk4MjYgKiByZWxBemltX25vcm07CgogIHJldHVybiB0YW5zaWcoc3VtKTsKfQoKZnVuY3Rpb24gbmV1cm9uNChiMDNfbm9ybSxiMDRfbm9ybSxiMDVfbm9ybSxiMDZfbm9ybSxiMDdfbm9ybSxiOGFfbm9ybSxiMTFfbm9ybSxiMTJfbm9ybSwgdmlld1plbl9ub3JtLHN1blplbl9ub3JtLHJlbEF6aW1fbm9ybSkgewogIHZhciBzdW0gPQoJKyAxLjUzMzk4ODI2NDY1NTQyMAoJLSAwLjEwOTM2NjU5MzY3MDQwNCAqIGIwM19ub3JtCgktIDAuMDcxMDQ2MjYyOTcyNzI5ICogYjA0X25vcm0KCSsgMC4wNjQ1ODI0MTE0NzgzMjAgKiBiMDVfbm9ybQoJKyAyLjkwNjMyNTIzNjgyMzE2MCAqIGIwNl9ub3JtCgktIDAuNjczODczMTA4OTc5MTYzICogYjA3X25vcm0KCS0gMy44MzgwNTE4NjgyODA4NDAgKiBiOGFfbm9ybQoJKyAxLjY5NTk3OTM0NDUzMTUzMCAqIGIxMV9ub3JtCgkrIDAuMDQ2OTUwMjk2MDgxNzEzICogYjEyX25vcm0KCS0gMC4wNDk3MDk2NTI2ODgzNjUgKiB2aWV3WmVuX25vcm0KCSsgMC4wMjE4Mjk1NDU0MzA5OTQgKiBzdW5aZW5fbm9ybQoJKyAwLjA1NzQ4MzgyNzEwNDA5MSAqIHJlbEF6aW1fbm9ybTsKCiAgcmV0dXJuIHRhbnNpZyhzdW0pOwp9CgpmdW5jdGlvbiBuZXVyb241KGIwM19ub3JtLGIwNF9ub3JtLGIwNV9ub3JtLGIwNl9ub3JtLGIwN19ub3JtLGI4YV9ub3JtLGIxMV9ub3JtLGIxMl9ub3JtLCB2aWV3WmVuX25vcm0sc3VuWmVuX25vcm0scmVsQXppbV9ub3JtKSB7CiAgdmFyIHN1bSA9CgkrIDMuMDI0MTE1OTMwNzU3MjMwCgktIDAuMDg5OTM5NDE2MTU5OTY5ICogYjAzX25vcm0KCSsgMC4xNzUzOTU0ODMxMDYxNDcgKiBiMDRfbm9ybQoJLSAwLjA4MTg0NzMyOTE3MjYyMCAqIGIwNV9ub3JtCgkrIDIuMjE5ODk1MzY3NDg3NzkwICogYjA2X25vcm0KCSsgMS43MTM4NzM5NzUxMzY4NTAgKiBiMDdfbm9ybQoJKyAwLjcxMzA2OTE4NjA5OTUzNCAqIGI4YV9ub3JtCgkrIDAuMTM4OTcwODEzNDk5MjAxICogYjExX25vcm0KCS0gMC4wNjA3NzE3NjE1MTgwMjUgKiBiMTJfbm9ybQoJKyAwLjEyNDI2MzM0MTI1NTQ3MyAqIHZpZXdaZW5fbm9ybQoJKyAwLjIxMDA4NjE0MDQwNDM1MSAqIHN1blplbl9ub3JtCgktIDAuMTgzODc4MTM4NzAwMzQxICogcmVsQXppbV9ub3JtOwoKICByZXR1cm4gdGFuc2lnKHN1bSk7Cn0KCmZ1bmN0aW9uIGxheWVyMihuZXVyb24xLCBuZXVyb24yLCBuZXVyb24zLCBuZXVyb240LCBuZXVyb241KSB7CiAgdmFyIHN1bSA9CgkrIDEuMDk2OTYzMTA3MDc3MjIwCgktIDEuNTAwMTM1NDg5NzI4NzMwICogbmV1cm9uMQoJLSAwLjA5NjI4MzI2OTEyMTUwMyAqIG5ldXJvbjIKCS0gMC4xOTQ5MzU5MzA1NzcwOTQgKiBuZXVyb24zCgktIDAuMzUyMzA1ODk1NzU1NTkxICogbmV1cm9uNAoJKyAwLjA3NTEwNzQxNTg0NzQ3MyAqIG5ldXJvbjU7CgogIHJldHVybiBzdW07Cn0KCmZ1bmN0aW9uIG5vcm1hbGl6ZSh1bm5vcm1hbGl6ZWQsIG1pbiwgbWF4KSB7CiAgcmV0dXJuIDIgKiAodW5ub3JtYWxpemVkIC0gbWluKSAvIChtYXggLSBtaW4pIC0gMTsKfQpmdW5jdGlvbiBkZW5vcm1hbGl6ZShub3JtYWxpemVkLCBtaW4sIG1heCkgewogIHJldHVybiAwLjUgKiAobm9ybWFsaXplZCArIDEpICogKG1heCAtIG1pbikgKyBtaW47Cn0KZnVuY3Rpb24gdGFuc2lnKGlucHV0KSB7CiAgcmV0dXJuIDIgLyAoMSArIE1hdGguZXhwKC0yICogaW5wdXQpKSAtIDE7IAp9CgpmdW5jdGlvbiBzZXR1cCgpIHsKICByZXR1cm4gewogICAgaW5wdXQ6IFt7CiAgICAgIGJhbmRzOiBbCiAgICAgICAgICAiQjAzIiwKICAgICAgICAgICJCMDQiLAogICAgICAgICAgIkIwNSIsCiAgICAgICAgICAiQjA2IiwKICAgICAgICAgICJCMDciLAogICAgICAgICAgIkI4QSIsCiAgICAgICAgICAiQjExIiwKICAgICAgICAgICJCMTIiLAogICAgICAgICAgInZpZXdaZW5pdGhNZWFuIiwKICAgICAgICAgICJ2aWV3QXppbXV0aE1lYW4iLAogICAgICAgICAgInN1blplbml0aEFuZ2xlcyIsCiAgICAgICAgICAic3VuQXppbXV0aEFuZ2xlcyIKICAgICAgXQogICAgfV0sCiAgICBvdXRwdXQ6IFsKICAgICAgICB7CiAgICAgICAgICBpZDogImRlZmF1bHQiLAogICAgICAgICAgc2FtcGxlVHlwZTogIkFVVE8iLAogICAgICAgICAgYmFuZHM6IDEKICAgICAgICB9CiAgICBdCiAgfQp9CgpmdW5jdGlvbiBldmFsdWF0ZVBpeGVsKHNhbXBsZSwgc2NlbmUsIG1ldGFkYXRhLCBjdXN0b21EYXRhLCBvdXRwdXRNZXRhZGF0YSkgewogIGNvbnN0IHJlc3VsdCA9IGV2YWx1YXRlUGl4ZWxPcmlnKFtzYW1wbGVdLCBbc2NlbmVdLCBtZXRhZGF0YSwgY3VzdG9tRGF0YSwgb3V0cHV0TWV0YWRhdGEpOwogIHJldHVybiByZXN1bHRbT2JqZWN0LmtleXMocmVzdWx0KVswXV07Cn0K&evalscripturl=https://raw.githubusercontent.com/sentinel-hub/custom-scripts/master/sentinel-2/lai/script.js
